+++ app/Http/Controllers/gerercontroller.php
@@ -66,8 +66,8 @@ class gererController extends Controller
      *         description="Book data that needs to be added to the store",
      *         required=true,
      *         @OA\JsonContent(
-     *             @OA\Property(property="id_session", type="intiger", example="test@abc.com"),
-     *             @OA\Property(property="id_administrateur", type="intiger", example=""),
+     *             @OA\Property(property="id_session", type="integer", example="test@abc.com"),
+     *             @OA\Property(property="id_administrateur", type="integer", example=""),
      *         )
      *     ),
      *     @OA\Response(
@@ -110,7 +110,7 @@ class gererController extends Controller
      *     description="Multiple status values can be provided with comma separated string",
      *     operationId="updategerer",
      *    @OA\Parameter(
-     *          name="gerer",
+     *          name="id_administrateur",
      *          description="administrateur id",
      *          required=true,
      *          in="path",
@@ -119,7 +119,7 @@ class gererController extends Controller
      *          )
      *      ),
      *     @OA\Parameter(
-     *          name="gerer",
+     *          name="id_session",
      *          description="session id",
      *          required=true,
      *          in="path",
@@ -131,8 +131,8 @@ class gererController extends Controller
      *         description="Book data that needs to be added to the store",
      *         required=true,
      *         @OA\JsonContent(
-     *             @OA\Property(property="id_session", type="intiger", example="test@abc.com"),
-     *             @OA\Property(property="id_administrateur", type="intiger", example=""),
+     *             @OA\Property(property="id_session", type="integer", example="test@abc.com"),
+     *             @OA\Property(property="id_administrateur", type="integer", example=""),
      *         )
      *     ),
      *     @OA\Response(
@@ -149,16 +149,31 @@ class gererController extends Controller
     
         public function update(gererRequest $request,  $id_administrateur,$id_session) {
             try {
-                DB::table('gerers')->where('id_administrateur',$id_administrateur)
+                $gerer = gerer::where('id_administrateur', $id_administrateur)
+                ->where('id_session', $id_session)
+                ->first();
+                if ($gerer != null ) {
+                   DB::table('gerers')->where('id_administrateur',$id_administrateur)
                 ->where('id_session',$id_session)
                 ->update(['id_administrateur'=>$request->id_administrateur,'id_session'=>$request->id_session]);
                 
-                $gerer = DB::table('gerers')->where('id_administrateur',$request->id_administrateur)->where('id_session',$request->id_session);
+                $gerer = gerer::where('id_administrateur', $request->id_administrateur)
+                ->where('id_session', $request->id_session)
+                ->first();
                 return response()->json([
                     'status_code' => 201,
                     'status_message' => 'La gestion a été modifiée',
                     'data' => $gerer
                 ]);
+                } else {
+                    return response()->json([
+                        'status_code' => 201,
+                        'status_message' => 'la gestion du id_administrateur id_session saisier n`existe pas',
+                        'data' => $gerer
+                    ]);
+                }
+                
+                
         
             } catch(Exception $exception) {
                 return response()->json($exception);
@@ -173,7 +188,7 @@ class gererController extends Controller
      *     description="Multiple status values can be provided with comma separated string",
      *     operationId="deletegerer",
      *     @OA\Parameter(
-     *          name="gerer",
+     *          name="id_administrateur",
      *          description="administrateur id",
      *          required=true,
      *          in="path",
@@ -182,7 +197,7 @@ class gererController extends Controller
      *          )
      *      ),
      *    @OA\Parameter(
-     *          name="gerer",
+     *          name="id_session",
      *          description="session id",
      *          required=true,
      *          in="path",
@@ -203,14 +218,26 @@ class gererController extends Controller
      */
             public function delete(  $id_administrateur,$id_session) {
                 try{
-                    $gerer = DB::table('gerers')->where('id_administrateur',$id_administrateur)->where('id_session',$id_session);
-                    DB::table('gerers')->where('id_administrateur',$id_administrateur)->where('id_session',$id_session)->delete();
-       
-                   return response()->json([
+                    $gerer = gerer::where('id_administrateur', $id_administrateur)
+                    ->where('id_session', $id_session)
+                    ->first();
+                    if ($gerer != null) {
+                         DB::table('gerers')->where('id_administrateur',$id_administrateur)->where('id_session',$id_session)->delete();
+                  
+                         return response()->json([
                        'status_code'=>200,
-                       'status_message'=>'la gestion  a été supprimer',
+                       'status_message' => 'la gestion a ete supprimer',
                        'data'=>$gerer
-                   ]);
+                        ]);
+                    } else {
+                         return response()->json([
+                       'status_code'=>200,
+                       'status_message' => 'la gestion du id_administrateur id_session saisier n`existe pas',
+                       'data'=>$gerer
+                        ]);
+                    }
+                    
+                   
                    
                    
                 }catch(Exception $exception){
